requirements.txt
python-decouple==3.8
selenium==4.17.2
